# swagger.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Meu Amigo PET",
		"version": "1.0.0",
		"description": "Meu Amigo PET é uma API REST cujo objetivo principal é conectar usuários, sejam eles pais de PETs ou aspirantes ao cargo, com objetivos semelhantes: adoção ou cruzamento. A conexão entre usuários se dá por meio de anúncios, os quais podem ser de adoção ou cruzamento. Assim, usuários logados podem acessar o 'feed de anúnicos' e obter informações de pets e usuários. A API visa atender usuários comuns e administradores com funcionalidades diferentes para cada um dos tipos de contas. Os usuários devem estar logados no servidor para acessar as diferentes rotas e suas funcionalidades. A aplicação foi construida usando-se Node.JS, módulos NPM (específicados no arquivo **package.json**), o framework Express e o [GitHub](https: //github.com/Alisson3366/projeto-de-back-end)."
	},
	"servers": [
		{
			"url": "http://localhost:8080"
		}
	],
...
	"paths": {
...
		"/usuarios": {
			"get": {
				"tags": ["usuarios"],
				"summary": "Consulta de todos os usuários",
				"description": "Consulta de todos os usuários",
				"operationId": "consultaUsuarios",
				"responses": {
					"200": {
						"description": "Consulta efetuada com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/USUARIOS"
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/usuario/{id}": {
			"get": {
				"tags": ["usuarios"],
				"summary": "Consulta de usuário específico pelo ID",
				"description": "Consulta de usuário específico pelo ID",
				"operationId": "consultaUsuarioId",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "ID do usuário a ser consultado.",
						"required": true,
						"schema": {
							"type": "string",
							"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Consulta efetuada com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/USUARIOS"
								}
							}
						}
					},
					"400": {
						"description": "ID de usuário inválido!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Token não informado ou inválido. Acesso negado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Usuário não localizado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			},
			"patch": {
				"tags": ["usuarios"],
				"summary": "Atualiza um usuário já existente",
				"description": "Atualiza um usuário já existente",
				"operationId": "atualizaUsuario",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "ID do usuário a ser atualizado.",
						"required": true,
						"schema": {
							"type": "string",
							"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
						}
					}
				],
				"requestBody": {
					"description": "Atualiza um usuário já existente.",
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string",
										"example": "user7@mail.com"
									},
									"senha": {
										"type": "string",
										"example": "teste7"
									},
									"nome": {
										"type": "string",
										"example": "JOAQUIM PEREIRA"
									}
								}
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Usuário atualizado com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Mensagem": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "ID de usuário inválido!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Token não informado ou inválido. Acesso negado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Usuário não localizado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Nome, email ou senha inválidos!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["usuarios"],
				"summary": "Deleta um usuário da API",
				"description": "Deleta um usuário da API",
				"operationId": "deletaUsuario",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "ID do usuário a ser deletado.",
						"required": true,
						"schema": {
							"type": "string",
							"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Usuário deletado com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Mensagem": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "ID de usuário inválido!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Token não informado ou inválido. Acesso negado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Usuário não localizado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/anuncio/usuario/{id}": {
			"patch": {
				"tags": ["anuncios"],
				"summary": "Atualiza um anúncio já existente de um usuário",
				"description": "Atualiza um anúncio já existente de um usuário",
				"operationId": "atualizaAnuncioUsuario",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "ID do anúncio a ser atualizado.",
						"required": true,
						"schema": {
							"type": "string",
							"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
						}
					}
				],
				"requestBody": {
					"description": "Atualiza um anúncio já existente de um usuário",
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"tipo": {
										"type": "string",
										"example": "cruzamento"
									},
									"titulo": {
										"type": "string",
										"example": "PROCURO UM PARCEIRO DE CRUZA PARA A MINHA BORDER COLLIE"
									},
									"sexo": {
										"type": "string",
										"example": "fêmea"
									},
									"raca": {
										"type": "string",
										"example": "border collie"
									},
									"quantidade": {
										"type": "integer",
										"example": 0
									}
								}
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Anúncio atualizado com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Mensagem": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "ID de anúncio inválido!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Token não informado ou inválido. Acesso negado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Anúncio não localizado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Tipo, titulo, sexo, raca ou quantidade inválidos!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["anuncios"],
				"summary": "Deleta o anúncio de um usuário da API",
				"description": "Deleta o anúncio de um usuário da API",
				"operationId": "deletaAnuncioUsuario",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "ID do anúncio a ser deletado.",
						"required": true,
						"schema": {
							"type": "string",
							"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Anúncio deletado com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Mensagem": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "ID de anúncio inválido!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Token não informado ou inválido. Acesso negado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Anúncio não localizado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/pet/usuario/{id}": {
			"patch": {
				"tags": ["pets"],
				"summary": "Atualiza um pet já existente de um usuário",
				"description": "Atualiza um pet já existente de um usuário",
				"operationId": "atualizaPetUsuario",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "ID do pet a ser atualizado.",
						"required": true,
						"schema": {
							"type": "string",
							"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
						}
					}
				],
				"requestBody": {
					"description": "Atualiza um pet já existente de um usuário",
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"nome": {
										"type": "string",
										"example": "PITOCO"
									},
									"sexo": {
										"type": "string",
										"example": "fêmea"
									},
									"raca": {
										"type": "string",
										"example": "srd"
									}
								}
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Pet atualizado com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Mensagem": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "ID de pet inválido!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Token não informado ou inválido. Acesso negado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Pet não localizado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Nome, sexo ou raca inválidos!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["pets"],
				"summary": "Deleta o pet de um usuário da API",
				"description": "Deleta o pet de um usuário da API",
				"operationId": "deletaPetUsuario",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "ID do pet a ser deletado.",
						"required": true,
						"schema": {
							"type": "string",
							"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Pet deletado com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Mensagem": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "ID de pet inválido!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Token não informado ou inválido. Acesso negado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Pet não localizado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"USUARIOS": {
				"required": ["email", "senha", "nome"],
				"type": "object",
				"properties": {
					"_id": {
						"type": "string",
						"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
					},
					"email": {
						"type": "string",
						"example": "user1@mail.com"
					},
					"senha": {
						"type": "string",
						"example": "teste1"
					},
					"nome": {
						"type": "string",
						"example": "ARNALDO SANTIAGO"
					},
					"pets": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/PETS"
						}
					},
					"anuncios": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ANUNCIOS"
						}
					}
				}
			},
			"ANUNCIOS": {
				"required": ["tipo", "titulo", "sexo", "raca"],
				"type": "object",
				"properties": {
					"_id": {
						"type": "string",
						"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
					},
					"donoAnuncio": {
						"type": "string",
						"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
					},
					"tipo": {
						"type": "string",
						"example": "adoção"
					},
					"titulo": {
						"type": "string",
						"example": "3 FILHOTES DE PIT BUL PARA ADOÇÃO"
					},
					"sexo": {
						"type": "string",
						"example": "ambos"
					},
					"raca": {
						"type": "string",
						"example": "american pit bull terrier"
					},
					"quantidade": {
						"type": "integer",
						"example": 4
					}
				}
			},
			"PETS": {
				"required": ["nome", "sexo", "raca"],
				"type": "object",
				"properties": {
					"_id": {
						"type": "string",
						"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
					},
					"donoPet": {
						"type": "string",
						"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
					},
					"nome": {
						"type": "string",
						"example": "NINA"
					},
					"sexo": {
						"type": "string",
						"example": "fêmea"
					},
					"raca": {
						"type": "string",
						"example": "beagle"
					}
				}
			}
		}
	}
}